Teto de nivel 255: os tres defeitos que o proprio teto criou
MAX_LEVEL ja valia 255 e as tabelas de EXP ja tinham as 256 entradas, mas
ninguem tinha auditado o que o motor faz acima de 100. Os tres achados sao
consequencia direta do teto novo e nao existiam com 100:

- TryIncrementMonLevel somava 1 num u8: 255 + 1 dava 0, a guarda
  `nextLevel > GetCurrentLevelCap()` virava `0 > 255` e o SetMonData
  gravava NIVEL 0 no Pokemon do teto. Hoje so o daycare chama, e so nao
  corrompe porque o CalculateMonStats seguinte recalcula pela EXP.
- moveDamage e passiveHpUpdate eram s16. O termo de nivel do dano,
  (2 * nivel / 5 + 2), sai de 42 para 104: um golpe forte com STAB e
  fraqueza passa de 63 mil, enrolava para negativo e CURAVA o alvo.
  Viraram s32; e struct de EWRAM, nao entra em save, e custa 16 bytes.
- Tres leituras gExperienceTables[...][level + 1] sem min(): no teto elas
  leem a linha do growth rate seguinte, e no ultimo (Slow) saem do array.

Nivel nao esta em save: struct BoxPokemon nao tem campo de nivel, ele e
derivado de experience:26 (52.728.772 no nivel 255 contra teto de
67.108.863, 21% de folga), e o u8 level de struct Pokemon e cache, com
255 sendo exatamente o maximo do tipo. guarda_save.py verde.

+++ include/battle.h
@@ -697,8 +697,14 @@ struct BattleStruct
     u8 pursuitStoredSwitch; // Stored id for the Pursuit target's switch
     s32 battlerExpReward;
     enum Species prevTurnSpecies[MAX_BATTLERS_COUNT]; // Stores species the AI has in play at start of turn
-    s16 passiveHpUpdate[MAX_BATTLERS_COUNT]; // non-move damage and healing
-    s16 moveDamage[MAX_BATTLERS_COUNT];
+    // ponytail: s32, e nao os s16 do upstream, porque MAX_LEVEL aqui e 255.
+    // O termo de nivel do dano, (2 * nivel / 5 + 2), sai de 42 no nivel 100
+    // para 104 no 255: o dano de um golpe forte com STAB e fraqueza passa de
+    // 63 mil, e em s16 ele enrolava para negativo e CURAVA o alvo. Nao esta em
+    // save nenhum (BattleStruct e de EWRAM, so vive dentro da batalha), custa
+    // 16 bytes de EWRAM, e a batalha nunca serializa a struct por offset.
+    s32 passiveHpUpdate[MAX_BATTLERS_COUNT]; // non-move damage and healing
+    s32 moveDamage[MAX_BATTLERS_COUNT];
     u16 innardsOutHpLost[MAX_BATTLERS_COUNT];
     u32 moveResultFlags[MAX_BATTLERS_COUNT];
     u32 savedMoveResultFlags[MAX_BATTLERS_COUNT]; // for Bounced moves
